+++ prisma/migrations/20260222125343_add_delivered_status/migration.sql
@@ -1,0 +1,5 @@
+-- AlterEnum
+ALTER TYPE "OrderStatus" ADD VALUE 'DELIVERED';
+
+-- AlterTable
+ALTER TABLE "Order" ADD COLUMN     "deliveredAt" TIMESTAMP(3);